refactor(usage): tighten module boundary and address review feedback
- usage/index.ts: strip composition-root exports per module-boundary
  rule (matches approvals, egress-rules)
- compose.ts: fix import path after bootstrap-agents move to services/
- services/bootstrap-agents.ts: fix domain types import path
- routes.ts: replace per-handler gate() calls with a router-scoped
  inspector-role middleware
- domain/types.ts: inline outcome union, drop events.ts import
- docs: bump channels.md Last verified, document ChannelTurnRelayed
  emission shape; clarify platform-topology protocol rows
