+++ packages/api-server/src/modules/usage/services/bootstrap-agents.ts
@@ -1,4 +1,4 @@
-import type { AgentRegistryRow } from "./domain/types.js";
+import type { AgentRegistryRow } from "../domain/types.js";
 
 export type BootstrapAgentsDeps = {
   listIdentities: () => Promise<{ id: string; owner: string }[]>;
